Console now outputs API version.
Seems a bit pointless to display the version of the Terraria Server
we're on because that will probably never change again. On the contrary,
the server doesn't yet have an updater, nor does it display which
version it is. This is kind of important if tweaks are done to the
server and someone doesn't have the latest version.

We can also tick the revision & build numbers without tripping the up to
date check, so stop not versioning minor changes!