Completely overhauled sendq
The new system now uses block allocators.  It is lock-free, and as a result
it is about 10,000x faster without the possibility of deadlocking.

It works on allocating small and large packets on separate packet heaps in
128-byte or 32k blocks for small and large respectively.  The sendq still
has a queue but not in the traditional sense as packets are allocated
out of a pool of free blocks atomically without having to synchronize access
to a managed collection.  Packets are enlisted in the same way by atomically
flicking a queue bit on in the pool, and the client's write thread will take
care of sending any packet with the bit on and releasing the packet back
into the pool once its done sending.

More than one packet may be sent in one iteration.

Performance good with 55+ players @ time of commit. Fixes deadlocking