SendQ sequencing mechanism
The sequencing mechanism works by modifying SendData slightly so that it may
add packets to a sequence linked list, which may consist of small and large
heap items, and are sent in order in one batch.

This fixes the tile and chest glitch by sending components in each in
order they are required.

Fixes chest glitches
Fixes rotating tile glitches