Fix issue #1056
So, basically. When I added PlayerTriggerPressurePlate hook, I added
SwitchTile back into Update() , in player class. Which caused HitSwitch
to trigger twice. (Once in MessageBuffer, once in SwitchTile).  I
thought the best solution would be to simply modify messagebuffer not to
trigger HitSwitch if its a pressureplate. This way the hook stays fully
functional. And the double trigger issue with pressure plates is fixed.