Prevented SendQ from trying to send non-allocated items
The disconnection issue is caused by the send queue trying to
send heap items that aren't allocated as AllocAndSet will
call enqueue if the allocation had failed, disconnecting the
client.

Also placed a barrier around the nullification of the heap
when the write thread has exited, which will stop the broadcasters
from trying to write to a null heap, but *may* introduce a
deadlock as there is now a second syncRoot.  Will need some
mad testing.

Fixes NyxStudios/TShock#1086, NyxStudios/TShock#1076

*NEEDS TEST*