Added protection around GetData
Some mallicious packets are getting past CheckBytes and crashing
the game mechanism.  This patch includes protection for the
GetData stack, enabling the server to at least keep going on the
receipt of a corrupt or otherwise invalid piece of information
from the client.

The server will complain if an exception happens in GetData, with
a hexadecimal representation of the payload so that the error
may be artificially reproduced if the warning is reported.

The important thing is, corrupted data should not bring down a
server anymore.  Should resolve all sorts of overflow, and out
of bounds crashes.