SendQ now uses static buffers and a very tidy cleanup mechansm.
Write threads now look after themselves by polling WinSock checking
for dead sockets if there is nothing in the queue.  If a socket is
dead then the write thread is cleaned up.

As packet buffers are now static, there is no additional exchanging
of heaps to null, no allocRoot locking and the process is now
extremely simple.

Fixes 'Found server' issues, thread-locking, and possibly SSC
saving because of SocketReset not being called on dead clients.