Fixed traps not working in custom builds
The IL emitter seems to change around some Vector2 initializers
in later versions of mono and the Roslyn compilers in such a
way that the Vector2 coordinates were being set properly in the
constructors but being overwritten with zero in the auto
initializer.

Because of this problem, Wiring was causing NewProjectile calls
from traps and pressure plates being called at or near 0,0 for
the X and Y positions.  The projectiles were being sent just
with invalid coordinates.

This patch changes the offending code to use standard incrementers
which guarantee the modification of value types on the Vector2
struct after it has been fully initialized on the stack.

Result: Traps work again in custom TerrariaServerAPI builds. This
fixes #43.

Signed-off-by: Tyler Watson <tyler@tw.id.au>