Removed ThreadPool threads in favour of plain threads.
ThreadPools are optimised for short units of work and are not designed
to house long running operations or looped work.  In fact, deadlocks
may arise from use of Thread.Sleep() inside a ThreadPool thread and is
quite dangerous.

Running TShock on mono >= 4.2.3.4 requires this change, as the
ThreadPool is exhausted before the server even begins, leading to hung
processes.

Requires more testing, but should fix NyxStudios/TShock#1172.