Enable proper UTF-8 Console support on TSAPI
This patch introduces a `TShock.Modifications.UnicodeInput` modification that overrides the Re-Logic standard console initialization function which sets the Console I/O encoding to `Unicode`. Unicode is an alias for UTF-16LE which seems to be only supported on Windows 10's new console interface.  UTF-8 should cover enough of the codepages for the languages that Terraria supports via its -lang parameter, and doesn't carry the silly Unicode artifacts present when you type anything on macOS/Linux.

Fixes an issue where typing in the console yielded garbage on 1.3.5 on macOS and Linux.