Possible fix to -world not working correctly
Basically you currently have to specify -world, -worldpath & -worldname for the server to generate a world and reuse it.
This commit should accept just -world and auto populate the other two variables. My initial thoughts were to only set it when they were blank but Main.WorldPath always defaults to My Games\Terraria, and also if you are specifying -world you are specifying the name, path and file all-in-one anyway.