Patch the critical, world corrupting bug reported by @bartico6.
Thanks to @QuiCM and @bartico6 for helping fix & report this,
respectively.

Exploit: Send a packet with a length really really high to the
server. The server will accept it and either deadlock, crash,
corrupt the world, crash clients, or quite a lot more.

This is a critical exploit.

The length 1000 was arbitrarily chosen and may need to be tuned.