Disallow loading two dangerous plugins
During plugin review, it was discovered that AutoRegister <= 1.2.0
contains code that automatically creates user accounts with predictable
passwords. The passwords are using low entropy material (the IP address
of the currently connecting client, the name of the current player, and
the UUID) and as a result can be easily obtained and forged. On public
servers who use AutoRegister, passwords for the saved accounts are
stored in plaintext in the database, because the character name, recent
IP, and UUID are all stored in the database.

Because this plugin generates passwords which defeat the security of
TShock, and which cannot be disabled simply by turning UUID login off,
all passwords generated on public servers using this plugin must be
changed.

This change to TSAPI prevents the server from booting if the vulnerable
plugin is present, in order to stop server admins from unwittingly
running a knowingly dangerous plugin.

Additionally, this change also prohibits loading another vulnerable
plugin. That plugin has not been identified yet, but will be released
after this update has been released.

The mechanism for disallowing loading of plugins relies on the TShock
team having evidence that a plugin is dangerous. It's not a tool that we
expect to use often. As a result, the capability to update it relies on
updating TSAPI, which is a fine compromise for now.

This commit is the combination of a few commits. Notably, the original
design of this check was mostly living in ServerApi. The goal was to
merge this really rapidly, and that didn't happen. As a result, more
time was taken to refactor it into a more modular component, given that
the urgency had decreased.

A huge thanks to people who reviewed this throughout its development.

Generalize dangerous plugins

This generalizes dangerous plugins to reject more than just one plugin.
Also checks at the point the assembly is loaded instead of after the
PluginContainer is created for good measure.

Fluff up comments and change access levels on MA

Add links to security advisories in comments

Sharpen methods a little for hashing things

Document the dangerousPlugins K,V

The goal here is to get us to the point where the codebase doesn't suck
(or so they say).

Co-authored-by: Zack Piispanen <pikakillers@hotmail.com>

Self contain dangerous plugins in init function

Move dangerous plugin check to separate object

Remove extra include in serverapi

Move logic for detecting mplugins around

Cleanup copypasta as a result of reorg

Continue copypasta cleanup

Look, I decided I wasn't going to do the submodule development thing
anymore. BUT the problem is that my entire pipeline requires it to
compile locally so I mean, here we are with these stupid commits.

Co-authored-by: Zack Piispanen <pikakillers@hotmail.com>