Update SSC.cs to target SendDataDirect correctly.
The dev team has added an additional parameter to SendData and SendDataDirect, which is unused and will be removed in the next hotfix. To avoid having to update all plugins - as they would look for the 11 parameter length version of SendData - we set SendData in OTAPI to internal to avoid ambiguity, and injected another.

For reasons (DeathCradle knows better) SendDataDirect also turns internal as we set SendData, so the hotfix for this, is to adjust a line in TSAPI SSC modification.